// carswitchFrontend/src/api/model/loggedInData.ts
/**
 * SwitchCar API
 * SwitchCar API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: contact@algebratec.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { LoggedInDataToken } from './loggedInDataToken';
import { User } from './user';

export interface LoggedInData { 
    user?: User;
    token?: LoggedInDataToken;
}